update code in both llm and gateway filter
